pyproject.toml
[project]
name = "pipecat-mistral-tts-demo"
version = "0.1.0"
description = "Demo reproducing choppy audio when streaming Mistral Voxtral TTS through Pipecat + Daily WebRTC"
requires-python = ">=3.11"
dependencies = [
    "pipecat-ai[daily,silero]==0.0.108",
    "mistralai>=2.2.0",
    "numpy",
    "aiohttp",
    "python-dotenv",
    "loguru",
    "matplotlib",
]
